Finished the first version of the site
Signe-off-by: Gabriel Faria <dev.gabriel.faria@gmail.com>

+++ backend/src/main/java/com/course/management/controller/CourseController.java
@@ -1,14 +1,16 @@
 package com.course.management.controller;
 
 import com.course.management.model.Course;
+import com.course.management.model.Student;
 import com.course.management.repository.CourseRepository;
+import com.course.management.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-
+import java.util.Optional;
 
 @RestController
 @CrossOrigin
@@ -16,6 +18,9 @@ public class CourseController {
 
     @Autowired
     private CourseRepository courseRepository;
+
+    @Autowired
+    private StudentRepository studentRepository;
 
     @GetMapping("/courses")
     public ResponseEntity<List<Course>> getCourses() {
@@ -28,7 +33,7 @@ public class CourseController {
     public ResponseEntity<Course> saveCourse(
             @RequestBody Course course) {
         Course savedCourse = courseRepository.save(course);
-        return new ResponseEntity<Course>(
+        return new ResponseEntity<>(
                 savedCourse,
                 HttpStatus.CREATED);
     }
@@ -36,9 +41,17 @@ public class CourseController {
     @DeleteMapping("/courses/{id}")
     public ResponseEntity<Void> removeCourse(
             @PathVariable("id") long id){
-        courseRepository.deleteById(id);
-        return new ResponseEntity<>(
-                HttpStatus.OK
-        );
+        Optional<Course> courseOptional = courseRepository.findById(id);
+        if (courseOptional.isPresent()) {
+            Course course = courseOptional.get();
+            List<Student> students = studentRepository.findByCourse(course);
+            if (!students.isEmpty()) {
+                return ResponseEntity.status(HttpStatus.CONFLICT).build();
+            }
+            courseRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
